raft.h: Add declarations for `struct raft_tracer` v2
Add new declaration for `struct raft_tracer` v2.

This new version of the tracer will support multiple types of events, such as
metrics-oriented events.

It also makes it possible to add new information to existing events without
breaking backward compatibilty. In particular, a new `level` field has ben added
to diagnostic messages, to be able to tweak the amount of messages being logged.

Note that the `raft_tracer->enabled` field of v1 is currently not strictly
needed, because:

1) if the user doesn't want tracing at all they should set `raft->tracer` to
   `NULL` (and that setting can be done conditionally by the user at any time,
   even dynamically after raft has started running)

2) if they want a tracer that can be dynamically enabled/disabled at runtime,
   the user-definable `raft_tracer->impl` field and/or the custom `raft_tracer->trace`
   function is what should be used to add custom logic (for example conditionally
   calling fprintf depending on an environment variable)

The v1 `struct raft_tracer` repurposes the enabled field with a version field,
with no loss in possible functionality for the user.

In C, the value of `false` is `0` and the value of `true` is `1`, so we can
detect that a `struct raft_tracer` object is using the v2 signature by checking
if `raft_tracer->version == 2`, because in that case the `enabled` field surely
wasn't set to `true` (regardless of the endianness of the system).

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>