benchmark/profiler: Use device start/end sector to filter commands
If the device being used for the benchmark is being used by the system for other
things (e.g. the device is a partition in the system disk), then spurious
commands might be caught by the profiler.

This commit makes the profiler ignore those commands, based on the start and end
sector.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>