raft.h: Add raft_step() definition
The raft_step() function will be the core of the v1 API, and it's used to drive
the raft engine forward.

It receives and event as input (such a timeout, a received message, an I/O
completion etc) and outputs a queue of tasks that consuming code should carry
out.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>