raft.h: Add configuration_last_snapshot_index field to struct raft
The new field will keep track of the log index of the last snapshot that was
taken, which will in turn be used to create RAFT_LOAD_SNAPSHOT tasks that
reference the most recent snapshot.

While not strictly necessary, including the target index in RAFT_LOAD_SNAPSHOT
tasks generally simplifies the consumer's implementation, which does not have to
figure out what the most recent snapshot is.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>