replication: Limit the number of entries included in AppendEntries
Up to now there was no limit in the number of entries that a leader would send
to a follower.

This introduces and hard-coded limit of 128, which at least prevents sending an
unbounded number of entries.

Further refinements could make this limit configurable and/or also take into
account the overall size of the entries being sent.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>