replication: Update r->commit index immediately
Don't lag r->commit_index unnecessarily. It is now updated immediately in a
single place, i.e. in replicationAppend(), which is called when receiving an
AppendEntries message.

Before that change it was updated immediately only if the AppendEntries message
was empty, while if the AppendEntries message was not empty, it would instead be
updated asynchronously after the raft_io->append_entries() request completes and
its callback is fired.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>