replication: Set a saturated status flag when entries are lagging
If a follower is not persisting entries fast enough (e.g. because the disk is
full), it will now stop accepting new entries, and will set a status flag for
informing the leader of this.

In turn, the leader will stop sending new entries until the follower clears the
flag.

Signed-off-by: Free Ekanayaka <free@ekanayaka.io>